requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.12.14
channels==4.2.0
charset-normalizer==3.4.1
class-registry==2.1.2
contourpy==1.3.1
crispy-bootstrap5==2024.10
cycler==0.12.1
Django==4.2.17
django-bootstrap-v5==1.0.11
django-crispy-forms==2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filters==1.3.2
fonttools==4.55.3
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
numpy==2.2.1
packaging==24.2
pillow==11.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.2
regex==2024.11.6
requests==2.32.3
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
urllib3==2.3.0
